ejemplos, enunciados, apuntes en borrador

+++ src/main/java/ar/unrn/patrones/estructurales/decorator/Obtenible.java
@@ -1,5 +1,7 @@
 package ar.unrn.patrones.estructurales.decorator;
 
+import java.math.BigInteger;
+
 public interface Obtenible {
-    int recuperar();
+    BigInteger recuperar();
 }
